+++ client/src/lib/socket.ts
@@ -3,12 +3,26 @@ import Logger from "./Logger";
 
 const url = process.env.REACT_APP_IS_DEV === "true" ? process.env.REACT_APP_SERVER_URL! : "";
 
-const socket = io(url);
+const socket = io(url, {
+  withCredentials: true,
+});
 const INTERVAL_1_MIN = 60_000; /* 1 minute interval */
 
 socket.on("connect", () => {
-  Logger.log("socket", `Connected to socket. ID: ${socket.id}`);
   socket.emit("CHECK_FOR_VERSION");
+});
+
+socket.on("connection_success", (e) => {
+  Logger.log("socket", `${e}. ID: ${socket.id}`);
+});
+
+socket.on("connection_error", (error) => {
+  if (error.status === "error") {
+    Logger.error("SOCKET", `Disconnected from socket. Error: ${error.error}`);
+
+    // Disconnect from socket if user is not authenticated
+    socket.close();
+  }
 });
 
 setInterval(() => {